LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/header-value-null-byte.any-expected.txt

CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/xhr/resources/parse-headers.py?my-custom-header=x%00x.

PASS Ensure fetch() rejects null bytes in headers
